# api.yaml
openapi: 3.0.3
info:
  title: RoboCore Server API
  version: 1.0.0
paths:
...
  /robots/{robot_id}:
    get:
      summary: Get robot configuration
      parameters:
        - in: path
          name: robot_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      summary: Update robot configuration
      parameters:
        - in: path
          name: robot_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Delete robot configuration
      parameters:
        - in: path
          name: robot_id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
servers:
  - url: https://api.robocore.dev/api/v1